+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Version 2022-dev
 -  update help of functions_common for rst (#650)
 -  switched topology map to using unique_ptr (#651)
 -  switch to using unique_ptr for cgmoleculedef (#652)
+-  better encapsulation to interaction, unique_ptr and const correctness (#658)
 
 Version 2021-rc.2 (released XX.01.21)
 =====================================